# microservices/books/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Biblioteca Digital API
  version: v1.0
  description: API Documentation for Biblioteca Digital
servers:
  - url: http://localhost:8083
    description: Módulo de Libros
...
paths:
...
  /books:
    post:
      tags: [Books]
      summary: Crear libro
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './models/openapi.yaml#/components/schemas/BookRequestDTO'
      responses:
        '201':
          description: Libro creado
          content:
            application/json:
              schema:
                $ref: './models/openapi.yaml#/components/schemas/BookResponseDTO'
  /books/page/{page}:
    get:
      tags: [Books]
      summary: Obtener todos los libros
      operationId: getAllBooks
      parameters:
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Lista de libros
          content:
            application/json:
              schema:
                $ref: './models/openapi.yaml#/components/schemas/PaginatedBookResponseDTO'
  /books/title/{title}/page/{page}:
    get:
      tags: [Books]
      summary: Obtener libros por título
      operationId: getBooksByTitle
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Lista de libros filtrados por título
          content:
            application/json:
              schema:
                $ref: './models/openapi.yaml#/components/schemas/PaginatedBookResponseDTO'
  /books/publication/{publicationId}/page/{page}:
    get:
      tags: [Books]
      summary: Obtener libros de una publicación
      operationId: getBooksByPublicationId
      parameters:
        - name: publicationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: ID de la publicación
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Lista de libros de una publicación
          content:
            application/json:
              schema:
                $ref: './models/openapi.yaml#/components/schemas/PaginatedBookResponseDTO'
  /books/author/{id}/page/{page}:
    get:
      tags: [Books]
      summary: Obtener los libros de un autor
      operationId: getBooksByAuthorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: page
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Lista de libros de un autor
          content:
            application/json:
              schema:
                $ref: './models/openapi.yaml#/components/schemas/PaginatedBookResponseDTO'
  /books/{id}:
    get:
      tags: [Books]
      summary: Obtener libro por ID
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Libro encontrado
          content:
            application/json:
              schema:
                $ref: './models/openapi.yaml#/components/schemas/BookResponseDTO'
    put:
      tags: [Books]
      summary: Actualizar libro
      operationId: updateBook
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './models/openapi.yaml#/components/schemas/BookRequestDTO'
      responses:
        '200':
          description: Libro actualizado
          content:
            application/json:
              schema:
                $ref: './models/openapi.yaml#/components/schemas/BookResponseDTO'
    delete:
      tags: [Books]
      summary: Eliminar libro (soft delete)
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Libro marcado como eliminado

  # ===========================
  # AUTHORS
  # ===========================
...